requirements.txt
blinker==1.6.2
cachelib==0.10.2
click==8.1.3
colorama==0.4.6
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.0
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.17
SQLAlchemy-Utils==0.41.1
typing_extensions==4.7.1
validators==0.20.0
Werkzeug==2.3.6
WTForms==3.0.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5